GTK: cleanup method names and ordering
