+++ druid-shell/src/keyboard.rs
@@ -531,6 +531,113 @@ impl From<i32> for KeyCode {
     }
 }
 
+#[cfg(target_os = "linux")]
+impl From<u32> for KeyCode {
+    #[allow(non_upper_case_globals)]
+    fn from(raw: u32) -> KeyCode {
+        use gdk::enums::key::*;
+        match raw {
+            a | A => KeyCode::KeyA,
+            s | S => KeyCode::KeyS,
+            d | D => KeyCode::KeyD,
+            f | F => KeyCode::KeyF,
+            h | H => KeyCode::KeyH,
+            g | G => KeyCode::KeyG,
+            z | Z => KeyCode::KeyZ,
+            x | X => KeyCode::KeyX,
+            c | C => KeyCode::KeyC,
+            v | V => KeyCode::KeyV,
+            b | B => KeyCode::KeyB,
+            q | Q => KeyCode::KeyQ,
+            w | W => KeyCode::KeyW,
+            e | E => KeyCode::KeyE,
+            r | R => KeyCode::KeyR,
+            y | Y => KeyCode::KeyY,
+            t | T => KeyCode::KeyT,
+            _1 => KeyCode::Key1,
+            _2 => KeyCode::Key2,
+            _3 => KeyCode::Key3,
+            _4 => KeyCode::Key4,
+            _6 => KeyCode::Key6,
+            _5 => KeyCode::Key5,
+            equal => KeyCode::Equals,
+            _9 => KeyCode::Key9,
+            _7 => KeyCode::Key7,
+            minus => KeyCode::Minus,
+            _8 => KeyCode::Key8,
+            _0 => KeyCode::Key0,
+            bracketright => KeyCode::RightBracket,
+            o | O => KeyCode::KeyO,
+            u | U => KeyCode::KeyU,
+            bracketleft => KeyCode::LeftBracket,
+            i | I => KeyCode::KeyI,
+            p | P => KeyCode::KeyP,
+            Return => KeyCode::Return,
+            l | L => KeyCode::KeyL,
+            j | J => KeyCode::KeyJ,
+            grave => KeyCode::Backtick,
+            k | K => KeyCode::KeyK,
+            semicolon => KeyCode::Semicolon,
+            backslash => KeyCode::Backslash,
+            comma => KeyCode::Comma,
+            slash => KeyCode::Slash,
+            n | N => KeyCode::KeyN,
+            m | M => KeyCode::KeyM,
+            period => KeyCode::Period,
+            Tab => KeyCode::Tab,
+            space => KeyCode::Space,
+            BackSpace => KeyCode::Backspace,
+            Escape => KeyCode::Escape,
+            Caps_Lock => KeyCode::CapsLock,
+            KP_Decimal => KeyCode::NumpadDecimal,
+            KP_Multiply => KeyCode::NumpadMultiply,
+            KP_Add => KeyCode::NumpadAdd,
+            Num_Lock => KeyCode::NumLock,
+            KP_Divide => KeyCode::NumpadDivide,
+            KP_Enter => KeyCode::NumpadEnter,
+            KP_Subtract => KeyCode::NumpadSubtract,
+            KP_Equal => KeyCode::NumpadEquals,
+            KP_0 => KeyCode::Numpad0,
+            KP_1 => KeyCode::Numpad1,
+            KP_2 => KeyCode::Numpad2,
+            KP_3 => KeyCode::Numpad3,
+            KP_4 => KeyCode::Numpad4,
+            KP_5 => KeyCode::Numpad5,
+            KP_6 => KeyCode::Numpad6,
+            KP_7 => KeyCode::Numpad7,
+            KP_8 => KeyCode::Numpad8,
+            KP_9 => KeyCode::Numpad9,
+            F5 => KeyCode::F5,
+            F6 => KeyCode::F6,
+            F7 => KeyCode::F7,
+            F3 => KeyCode::F3,
+            F8 => KeyCode::F8,
+            F9 => KeyCode::F9,
+            F10 => KeyCode::F10,
+            F11 => KeyCode::F11,
+            F12 => KeyCode::F12,
+            Insert => KeyCode::Insert,
+            Home => KeyCode::Home,
+            Page_Up => KeyCode::PageUp,
+            Delete => KeyCode::Delete,
+            F4 => KeyCode::F4,
+            End => KeyCode::End,
+            F2 => KeyCode::F2,
+            Page_Down => KeyCode::PageDown,
+            F1 => KeyCode::F1,
+            Left => KeyCode::ArrowLeft,
+            Right => KeyCode::ArrowRight,
+            Down => KeyCode::ArrowDown,
+            Up => KeyCode::ArrowUp,
+            quoteright => KeyCode::Quote,
+            _ => {
+                eprintln!("Warning: unknown keyval {}", raw);
+                KeyCode::Unknown(RawKeyCode::Linux(raw))
+            }
+        }
+    }
+}
+
 /// Should realistically be (8 * N) - 1; we need one byte for the length.
 const TINY_STR_CAPACITY: usize = 15;
